ctf2021_2022/w22w2/marvin-main/bot/requirements.txt
discord==1.7.3
mysql-connector-python
tabulate
